requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.1
billiard==3.6.3.0
boto3==1.16.56
botocore==1.19.56
bs4==0.0.1
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.7
certifi==2019.9.11
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cssselect2==0.3.0
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.6
django-admin-thumbnails==0.2.5
django-appconf==1.0.4
django-autocomplete-light==3.5.1
django-autoslug-iplweb==1.9.4.dev0
django-bootstrap4==1.1.1
django-ckeditor==5.9.0
django-cleanup==5.0.0
django-contactform==2.2.0
django-cors-headers==3.5.0
django-crispy-forms==1.8.1
django-currencies==0.10.1
django-heroku==0.3.1
django-imagekit==4.0.2
django-import-export==2.4.0
django-js-asset==1.2.2
django-jsonfield==1.4.0
django-model-search==2.2
django-modeltranslation==0.15.1
django-mptt==0.11.0
django-paypal==1.0.0
django-phonenumber-field==5.0.0
django-select2==7.4.2
django-storages==1.11.1
et-xmlfile==1.0.1
feedparser==6.0.2
filelock==3.0.12
flower==0.9.5
google-api-python-client==1.4.1
googletrans==3.0.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.8.25
html5lib==1.1
httpcore==0.9.1
httplib2==0.18.1
httpx==0.13.3
humanize==2.5.0
hyperframe==5.2.0
idna==2.8
importlib-metadata==1.7.0
jdcal==1.4.1
jmespath==0.10.0
joblib==0.16.0
jsonfield==2.1.1
kombu==4.6.11
ldif3==3.1.1
liqpay-python==1.0
MarkupPy==1.14
nltk==3.5
numpy==1.19.1
oauth2client==4.1.3
oauthlib==3.1.0
odfpy==1.4.1
opencv-python==4.4.0.42
openpyxl==3.0.5
phonenumbers==8.12.8
pilkit==2.0
Pillow==6.2.0
pipenv==2020.11.15
progressbar2==3.51.4
prometheus-client==0.8.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.0
Pyphen==0.9.5
python-card-me==0.9.3
python-dateutil==2.8.1
python-gettext==4.0
python-utils==2.4.0
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.3.1
redis==3.5.3
regex==2020.7.14
requests==2.22.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.7
s3transfer==0.3.4
sgmllib3k==1.0.0
six==1.15.0
sniffio==1.1.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.0
surlex==0.2.0
tablib==2.0.0
textblob==0.15.3
tinycss2==1.0.2
tornado==6.0.4
tqdm==4.48.2
unicodecsv==0.14.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.6
vine==1.3.0
virtualenv==20.3.1
virtualenv-clone==0.5.4
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
